refactor: rename core and ui modules to richtext and richtext-editor
