--- core/build.gradle.kts
@@ -1,12 +1,0 @@
-plugins {
-    id("arranger.android.library")
-}
-
-android {
-    namespace = "dev.mkeeda.arranger.core"
-}
-
-dependencies {
-    testImplementation(libs.junit)
-    testImplementation(libs.kotest.assertions.core)
-}
