--- core/src/main/kotlin/dev/mkeeda/arranger/core/text/RichSpan.kt
@@ -1,12 +1,0 @@
-package dev.mkeeda.arranger.core.text
-
-/**
- * Represents a span of attributes applied to a specific range within a [RichString].
- *
- * @property range The character range (inclusive) to which the attributes are applied.
- * @property attributes The [AttributeContainer] holding the attribute values for this range.
- */
-public data class RichSpan(
-    val range: IntRange,
-    val attributes: AttributeContainer,
-)
