+++ richtext/build.gradle.kts
@@ -1,0 +1,12 @@
+plugins {
+    id("arranger.android.library")
+}
+
+android {
+    namespace = "dev.mkeeda.arranger.richtext"
+}
+
+dependencies {
+    testImplementation(libs.junit)
+    testImplementation(libs.kotest.assertions.core)
+}
