+++ sample-app/build.gradle.kts
@@ -39,6 +39,6 @@ dependencies {
     implementation(libs.androidx.compose.ui.tooling.preview)
     debugImplementation(libs.androidx.compose.ui.tooling)
     implementation(libs.androidx.compose.material3)
-    implementation(project(":ui"))
-    implementation(project(":core"))
+    implementation(project(":richtext-editor"))
+    implementation(project(":richtext"))
 }
